fix: Move a NamespaceData test into its module
