+++ ingester/src/data/namespace.rs
@@ -312,22 +312,40 @@ impl<'a> Drop for ScopedSequenceNumber<'a> {
 
 #[cfg(test)]
 mod tests {
-    use std::{sync::Arc, time::Duration};
-
-    use data_types::{PartitionId, PartitionKey, ShardIndex};
-    use metric::{Attributes, Metric};
-
+    use super::*;
     use crate::{
         data::{
-            partition::{resolver::MockPartitionProvider, PartitionData, SortKeyState},
+            namespace::NamespaceData,
+            partition::{
+                resolver::{CatalogPartitionResolver, MockPartitionProvider},
+                PartitionData, SortKeyState,
+            },
             table::{name_resolver::mock::MockTableNameProvider, TableName},
         },
-        deferred_load,
-        lifecycle::mock_handle::MockLifecycleHandle,
+        deferred_load::{self, DeferredLoad},
+        lifecycle::{mock_handle::MockLifecycleHandle, LifecycleConfig, LifecycleManager},
         test_util::{make_write_op, TEST_TABLE},
     };
-
-    use super::*;
+    use assert_matches::assert_matches;
+    use data_types::{
+        ColumnId, ColumnSet, CompactionLevel, NamespaceSchema, ParquetFileParams, Sequence,
+        Timestamp,
+    };
+    use data_types::{PartitionId, PartitionKey, ShardIndex};
+    use dml::{DmlMeta, DmlWrite};
+    use hashbrown::HashMap;
+    use iox_catalog::{
+        interface::{Catalog, RepoCollection},
+        mem::MemCatalog,
+        validate_or_insert_schema,
+    };
+    use iox_time::{SystemProvider, Time};
+    use metric::{Attributes, Metric};
+    use metric::{MetricObserver, Observation};
+    use mutable_batch::MutableBatch;
+    use mutable_batch_lp::lines_to_batches;
+    use std::{ops::DerefMut, sync::Arc, time::Duration};
+    use uuid::Uuid;
 
     const SHARD_INDEX: ShardIndex = ShardIndex::new(24);
     const SHARD_ID: ShardId = ShardId::new(22);
@@ -407,5 +425,209 @@ mod tests {
         let name = ns.namespace_name().get().await;
         assert_eq!(&**name, NAMESPACE_NAME);
         assert_eq!(ns.namespace_name().to_string(), NAMESPACE_NAME);
+    }
+
+    #[tokio::test]
+    async fn buffer_operation_ignores_already_persisted_data() {
+        test_helpers::maybe_start_logging();
+        let metrics = Arc::new(metric::Registry::new());
+        let catalog: Arc<dyn Catalog> = Arc::new(MemCatalog::new(Arc::clone(&metrics)));
+        let mut repos = catalog.repositories().await;
+        let topic = repos.topics().create_or_get("whatevs").await.unwrap();
+        let query_pool = repos.query_pools().create_or_get("whatevs").await.unwrap();
+        let shard_index = ShardIndex::new(0);
+        let namespace = repos
+            .namespaces()
+            .create("foo", topic.id, query_pool.id)
+            .await
+            .unwrap();
+        let shard = repos
+            .shards()
+            .create_or_get(&topic, shard_index)
+            .await
+            .unwrap();
+
+        let schema = NamespaceSchema::new(namespace.id, topic.id, query_pool.id, 100);
+
+        let ignored_ts = Time::from_timestamp_millis(42).unwrap();
+
+        let batch = lines_to_batches("mem foo=1 10", 0).unwrap();
+        let w1 = DmlWrite::new(
+            namespace.id,
+            batch.clone(),
+            build_id_map(repos.deref_mut(), namespace.id, &batch).await,
+            "1970-01-01".into(),
+            DmlMeta::sequenced(
+                Sequence::new(ShardIndex::new(1), SequenceNumber::new(1)),
+                ignored_ts,
+                None,
+                50,
+            ),
+        );
+        let batch = lines_to_batches("mem foo=1 10", 0).unwrap();
+        let w2 = DmlWrite::new(
+            namespace.id,
+            batch.clone(),
+            build_id_map(repos.deref_mut(), namespace.id, &batch).await,
+            "1970-01-01".into(),
+            DmlMeta::sequenced(
+                Sequence::new(ShardIndex::new(1), SequenceNumber::new(2)),
+                ignored_ts,
+                None,
+                50,
+            ),
+        );
+
+        let _ = validate_or_insert_schema(w1.tables(), &schema, repos.deref_mut())
+            .await
+            .unwrap()
+            .unwrap();
+
+        // create some persisted state
+        let table = repos
+            .tables()
+            .create_or_get("mem", namespace.id)
+            .await
+            .unwrap();
+        let partition = repos
+            .partitions()
+            .create_or_get("1970-01-01".into(), shard.id, table.id)
+            .await
+            .unwrap();
+        repos
+            .partitions()
+            .update_persisted_sequence_number(partition.id, SequenceNumber::new(1))
+            .await
+            .unwrap();
+        let partition2 = repos
+            .partitions()
+            .create_or_get("1970-01-02".into(), shard.id, table.id)
+            .await
+            .unwrap();
+
+        let parquet_file_params = ParquetFileParams {
+            shard_id: shard.id,
+            namespace_id: namespace.id,
+            table_id: table.id,
+            partition_id: partition.id,
+            object_store_id: Uuid::new_v4(),
+            max_sequence_number: SequenceNumber::new(1),
+            min_time: Timestamp::new(1),
+            max_time: Timestamp::new(1),
+            file_size_bytes: 0,
+            row_count: 0,
+            compaction_level: CompactionLevel::Initial,
+            created_at: Timestamp::new(1),
+            column_set: ColumnSet::new([ColumnId::new(1), ColumnId::new(2)]),
+        };
+        repos
+            .parquet_files()
+            .create(parquet_file_params.clone())
+            .await
+            .unwrap();
+
+        // now create a parquet file in another partition with a much higher sequence persisted
+        // sequence number. We want to make sure that this doesn't cause our write in the other
+        // partition to get ignored.
+        let other_file_params = ParquetFileParams {
+            max_sequence_number: SequenceNumber::new(15),
+            object_store_id: Uuid::new_v4(),
+            partition_id: partition2.id,
+            ..parquet_file_params
+        };
+        repos
+            .parquet_files()
+            .create(other_file_params)
+            .await
+            .unwrap();
+        std::mem::drop(repos);
+
+        let manager = LifecycleManager::new(
+            LifecycleConfig::new(
+                1,
+                0,
+                0,
+                Duration::from_secs(1),
+                Duration::from_secs(1),
+                1000000,
+            ),
+            Arc::clone(&metrics),
+            Arc::new(SystemProvider::new()),
+        );
+
+        let partition_provider = Arc::new(CatalogPartitionResolver::new(Arc::clone(&catalog)));
+
+        let data = NamespaceData::new(
+            namespace.id,
+            DeferredLoad::new(Duration::from_millis(1), async { "foo".into() }),
+            Arc::new(MockTableNameProvider::new(table.name)),
+            shard.id,
+            partition_provider,
+            &metrics,
+        );
+
+        // w1 should be ignored because the per-partition replay offset is set
+        // to 1 already, so it shouldn't be buffered and the buffer should
+        // remain empty.
+        let action = data
+            .buffer_operation(DmlOperation::Write(w1), &manager.handle())
+            .await
+            .unwrap();
+        {
+            let table = data.table(table.id).unwrap();
+            assert!(table
+                .partitions()
+                .into_iter()
+                .all(|p| p.lock().get_query_data().is_none()));
+            assert_eq!(
+                table
+                    .get_partition_by_key(&"1970-01-01".into())
+                    .unwrap()
+                    .lock()
+                    .max_persisted_sequence_number(),
+                Some(SequenceNumber::new(1))
+            );
+        }
+        assert_matches!(action, DmlApplyAction::Skipped);
+
+        // w2 should be in the buffer
+        data.buffer_operation(DmlOperation::Write(w2), &manager.handle())
+            .await
+            .unwrap();
+
+        let table = data.table(table.id).unwrap();
+        let partition = table.get_partition_by_key(&"1970-01-01".into()).unwrap();
+        assert_eq!(
+            partition.lock().sequence_number_range().inclusive_min(),
+            Some(SequenceNumber::new(2))
+        );
+
+        assert_matches!(data.table_count().observe(), Observation::U64Counter(v) => {
+            assert_eq!(v, 1, "unexpected table count metric value");
+        });
+    }
+
+    pub async fn build_id_map<R>(
+        catalog: &mut R,
+        namespace_id: NamespaceId,
+        tables: &HashMap<String, MutableBatch>,
+    ) -> HashMap<String, TableId>
+    where
+        R: RepoCollection + ?Sized,
+    {
+        let mut ret = HashMap::with_capacity(tables.len());
+
+        for k in tables.keys() {
+            let id = catalog
+                .tables()
+                .create_or_get(k, namespace_id)
+                .await
+                .expect("table should create OK")
+                .id;
+
+            ret.insert(k.clone(), id);
+        }
+
+        ret
     }
 }
